Move sdk-connection and managed-by validator to shared
