+++ packages/back-end/src/api/sdk-connections/listSdkConnections.ts
@@ -1,3 +1,4 @@
+import { SDKConnectionInterface } from "shared/types/sdk-connection";
 import { ListSdkConnectionsResponse } from "back-end/types/openapi";
 import {
   findAllSDKConnectionsAcrossAllOrgs,
@@ -11,7 +12,6 @@ import {
   validateIsSuperUserRequest,
 } from "back-end/src/util/handler";
 import { listSdkConnectionsValidator } from "back-end/src/validators/openapi";
-import { SDKConnectionInterface } from "back-end/types/sdk-connection";
 
 export const listSdkConnections = createApiRequestHandler(
   listSdkConnectionsValidator,
