+++ packages/back-end/src/jobs/sdkWebhooks.ts
@@ -4,6 +4,7 @@ import md5 from "md5";
 import { getConnectionSDKCapabilities } from "shared/sdk-versioning";
 import { filterProjectsByEnvironmentWithNull } from "shared/util";
 import { Promise as BluebirdPromise } from "bluebird";
+import { SDKConnectionInterface } from "shared/types/sdk-connection";
 import { getFeatureDefinitions } from "back-end/src/services/features";
 import { WEBHOOKS } from "back-end/src/util/secrets";
 import { SDKPayloadKey } from "back-end/types/sdk-payload";
@@ -11,7 +12,6 @@ import {
   findSDKConnectionsByIds,
   findSDKConnectionsByOrganization,
 } from "back-end/src/models/SdkConnectionModel";
-import { SDKConnectionInterface } from "back-end/types/sdk-connection";
 import { logger } from "back-end/src/util/logger";
 import {
   findAllSdkWebhooksByConnection,
