+++ packages/back-end/src/models/SdkConnectionModel.ts
@@ -2,11 +2,7 @@ import mongoose from "mongoose";
 import uniqid from "uniqid";
 import { z } from "zod";
 import { isEqual, omit } from "lodash";
-import { ApiSdkConnection } from "back-end/types/openapi";
-import {
-  managedByValidator,
-  ManagedBy,
-} from "back-end/src/validators/managed-by";
+import { managedByValidator, ManagedBy } from "shared/validators";
 import {
   CreateSDKConnectionParams,
   EditSDKConnectionParams,
@@ -14,7 +10,8 @@ import {
   ProxyTestResult,
   SDKConnectionInterface,
   SDKLanguage,
-} from "back-end/types/sdk-connection";
+} from "shared/types/sdk-connection";
+import { ApiSdkConnection } from "back-end/types/openapi";
 import { cancellableFetch } from "back-end/src/util/http.util";
 import {
   IS_CLOUD,
