+++ packages/back-end/src/routers/experiment-template/template.validators.ts
@@ -1,9 +1,9 @@
 import { z } from "zod";
+import { statsEngines } from "shared/constants";
 import {
   featurePrerequisite,
   savedGroupTargeting,
 } from "back-end/src/validators/shared";
-import { statsEngines } from "back-end/src/util/constants";
 import { customMetricSlice } from "back-end/src/validators/experiments";
 
 export const experimentTemplateInterface = z
