+++ packages/back-end/src/routers/vercel-native-integration/vercel-native-integration.controller.ts
@@ -3,6 +3,7 @@ import { Request, Response } from "express";
 import { z } from "zod";
 import { createRemoteJWKSet, jwtVerify } from "jose";
 import { v4 as uuidv4 } from "uuid";
+import { ManagedBy } from "shared/validators";
 import {
   findVercelInstallationByInstallationId,
   VercelNativeIntegrationModel,
@@ -19,7 +20,6 @@ import {
   findOrganizationById,
 } from "back-end/src/models/OrganizationModel";
 import { createUser, getUserByEmail } from "back-end/src/models/UserModel";
-import { ManagedBy } from "back-end/src/validators/managed-by";
 import { ReqContextClass } from "back-end/src/services/context";
 import { ReqContext, OrganizationInterface } from "back-end/types/organization";
 import {
