--- packages/back-end/src/util/constants.ts
@@ -1,27 +1,0 @@
-export const sdkLanguages = [
-  "nocode-webflow",
-  "nocode-wordpress",
-  "nocode-shopify",
-  "nocode-other",
-  "javascript",
-  "nodejs",
-  "nextjs",
-  "react",
-  "php",
-  "ruby",
-  "python",
-  "go",
-  "java",
-  "csharp",
-  "android",
-  "ios",
-  "flutter",
-  "elixir",
-  "edge-cloudflare",
-  "edge-fastly",
-  "edge-lambda",
-  "edge-other",
-  "other",
-] as const;
-
-export const statsEngines = ["bayesian", "frequentist"] as const;
