+++ packages/back-end/src/validators/experiments.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
+import { statsEngines } from "shared/constants";
 import { windowTypeValidator } from "back-end/src/routers/fact-table/fact-table.validators";
-import { statsEngines } from "back-end/src/util/constants";
 import {
   namespaceValue,
   featurePrerequisite,
