+++ packages/back-end/src/validators/features.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { statsEngines } from "back-end/src/util/constants";
+import { statsEngines } from "shared/constants";
 import { safeRolloutStatusArray } from "back-end/src/validators/safe-rollout";
 import {
   featurePrerequisite,
