+++ packages/back-end/types/webhook.d.ts
@@ -1,4 +1,4 @@
-import { ManagedBy } from "back-end/src/validators/managed-by";
+import { ManagedBy } from "shared/validators";
 
 export interface WebhookInterface {
   id: string;
