+++ packages/front-end/components/Experiment/TabbedPage/ResultsTab.tsx
@@ -7,7 +7,7 @@ import {
 } from "back-end/types/report";
 import uniq from "lodash/uniq";
 import { VisualChangesetInterface } from "shared/types/visual-changeset";
-import { SDKConnectionInterface } from "back-end/types/sdk-connection";
+import { SDKConnectionInterface } from "shared/types/sdk-connection";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { DifferenceType } from "back-end/types/stats";
