+++ packages/front-end/components/Experiment/TabbedPage/SetupTabOverview.tsx
@@ -4,7 +4,7 @@ import {
 } from "back-end/types/experiment";
 import React, { useState } from "react";
 import { VisualChangesetInterface } from "shared/types/visual-changeset";
-import { SDKConnectionInterface } from "back-end/types/sdk-connection";
+import { SDKConnectionInterface } from "shared/types/sdk-connection";
 import Collapsible from "react-collapsible";
 import { FaAngleRight } from "react-icons/fa";
 import { Box, Flex, ScrollArea, Heading } from "@radix-ui/themes";
