+++ packages/front-end/components/Features/SDKConnections/SdkWebhooks.tsx
@@ -7,7 +7,7 @@ import {
   FaPaperPlane,
 } from "react-icons/fa";
 import { ago } from "shared/dates";
-import { SDKConnectionInterface } from "back-end/types/sdk-connection";
+import { SDKConnectionInterface } from "shared/types/sdk-connection";
 import useApi from "@/hooks/useApi";
 import EditSDKWebhooksModal, {
   CreateSDKWebhookModal,
