+++ packages/front-end/components/GuidedGetStarted/CheckSDKConnectionModal.tsx
@@ -1,4 +1,4 @@
-import { SDKConnectionInterface } from "back-end/types/sdk-connection";
+import { SDKConnectionInterface } from "shared/types/sdk-connection";
 import Modal from "@/components/Modal";
 import ConnectionDiagram from "@/components/Features/SDKConnections/ConnectionDiagram";
 import usePermissionsUtil from "@/hooks/usePermissionsUtils";
