+++ packages/front-end/components/InitialSetup/VerifyConnectionPage.tsx
@@ -1,4 +1,4 @@
-import { SDKConnectionInterface } from "back-end/types/sdk-connection";
+import { SDKConnectionInterface } from "shared/types/sdk-connection";
 import { useCallback, useEffect, useState } from "react";
 import {
   FaAngleDown,
