+++ packages/front-end/hooks/useSDKConnections.ts
@@ -1,4 +1,4 @@
-import { SDKConnectionInterface } from "back-end/types/sdk-connection";
+import { SDKConnectionInterface } from "shared/types/sdk-connection";
 import useApi from "./useApi";
 
 export default function useSDKConnections() {
