+++ packages/front-end/pages/setup/index.tsx
@@ -3,7 +3,7 @@ import {
   CreateSDKConnectionParams,
   SDKConnectionInterface,
   SDKLanguage,
-} from "back-end/types/sdk-connection";
+} from "shared/types/sdk-connection";
 import { getLatestSDKVersion, getSDKCapabilities } from "shared/sdk-versioning";
 import { ProjectInterface } from "back-end/types/project";
 import { Environment } from "back-end/types/organization";
