+++ packages/shared/src/constants.ts
@@ -94,3 +94,31 @@ export const SAFE_ROLLOUT_VARIATIONS = [
     weight: 0.5,
   },
 ];
+
+export const sdkLanguages = [
+  "nocode-webflow",
+  "nocode-wordpress",
+  "nocode-shopify",
+  "nocode-other",
+  "javascript",
+  "nodejs",
+  "nextjs",
+  "react",
+  "php",
+  "ruby",
+  "python",
+  "go",
+  "java",
+  "csharp",
+  "android",
+  "ios",
+  "flutter",
+  "elixir",
+  "edge-cloudflare",
+  "edge-fastly",
+  "edge-lambda",
+  "edge-other",
+  "other",
+] as const;
+
+export const statsEngines = ["bayesian", "frequentist"] as const;
