+++ packages/shared/src/index.ts
@@ -9,3 +9,4 @@ export * as types from "../types";
 export * as power from "./power";
 export * as health from "./health";
 export * as sql from "./sql";
+export * as validators from "./validators";
